requirements.txt
mutagen>=1.47.0
numpy>=2.1.2
parameterized>=0.9.0
python-ffmpeg>=2.0.12
secretstorage
yt-dlp>=2024.10.22